requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic==2.12.5
google-genai==1.47.0
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
python-docx==1.2.0
